app/requirements.txt
pandas==1.5.3
kafka-python==2.0.2    
requests
scikit-surprise==1.1.3 
tensorflow==2.8.0   
zeno
gunicorn
prometheus-flask-exporter
mlflow
Werkzeug>=2.0
matplotlib
prometheus_client==0.15.0
Flask==2.2.3
scikit-learn==1.1.3  
joblib==1.3.2
numpy==1.21.6
